Add CurriculumMentors compoment to Curriculum container
Refactor CurriculumMentors component.
Refactor MessagesBarService.
Add CurriculumMentors compoment to the Curriculum container.
