+++ eventually/static/src/components/curriculumMentors/CurriculumMentors.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import MentorItem from './MentorItem';
+import MentorItem from './CurriculumMentorItem';
 import Paper from 'material-ui/Paper';
 
 export default class CurriculumMentors extends React.Component {
@@ -15,7 +15,8 @@ export default class CurriculumMentors extends React.Component {
             <div style={this.props.style}>
                 <Paper zDepth={2}>
                     {this.props.mentors.map(mentor => (
-                        <MentorItem key={mentor.id.toString()}
+                        <MentorItem
+                            key={mentor.id.toString()}
                             id={mentor.id}
                             firstName={mentor.first_name}
                             lastName={mentor.last_name}
