+++ eventually/static/src/components/messagesBar/MessagesList.js
@@ -15,7 +15,8 @@ export default class MessagesList extends React.Component {
                 <MessagesSender />
                 <Paper zDepth={2}>
                     {this.props.messages.map(message => (
-                        <MessageItem key={message.id.toString()}
+                        <MessageItem
+                            key={message.id.toString()}
                             author={message.author}
                             avatar={message.avatar}
                             text={message.text}
