--- eventually/static/src/containers/curriculum/CurriculumService.js
@@ -1,29 +1,0 @@
-const mentorsService = () => {
-
-    let mentors = {
-        'mentors': [
-            {
-                'id': 1,
-                'first_name': 'John',
-                'last_name': 'Doe',
-                'avatar': 'johndoe'
-            },
-            {
-                'id': 2,
-                'first_name': 'Eric',
-                'last_name': 'Moreno',
-                'avatar': 'ericmoreno'
-            },
-            {
-                'id': 3,
-                'first_name': 'Mark',
-                'last_name': 'Smith',
-                'avatar': 'marksmith'
-            }
-        ]
-    };
-
-    return mentors;
-};
-
-export {mentorsService};
